+++ backend/models/stt.py
