--- backend_service/common/text_normalizer/en.py
@@ -1,16 +1,0 @@
-from transformers.models.whisper.english_normalizer import BasicTextNormalizer
-
-EN_BASIC_NORMALIZER = BasicTextNormalizer()
-
-def en_normalizer(text):
-    text = text.lower()
-    text = EN_BASIC_NORMALIZER(text)
-    text = text.strip()
-    return text
-
-if __name__ == "__main__":
-    import sys
-    sys.path.append('.')
-    text = ' hello world!!!! '
-    print(bytes(text, 'utf-8'))
-    print(bytes(en_normalizer(text), 'utf-8'))
